refactor: Extract footer data
